requirements.txt
absl-py==0.10.0
adabound==0.0.5
addict==2.4.0
aiohttp==3.8.1
aiosignal==1.2.0
albumentations==1.0.3
ale-py==0.7.3
antlr4-python3-runtime==4.8
apex @ file:///home/administrator/%E4%B8%8B%E8%BD%BD/apex
astor==0.8.1
astunparse==1.6.3
async-timeout==4.0.1
asynctest==0.13.0
attrs==21.2.0
autograd==1.3
ax-platform==0.1.9
backcall==0.2.0
botorch==0.2.1
cached-property==1.5.2
cachetools==4.1.1
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
charset-normalizer==2.0.7
ci-info==0.3.0
click==8.0.3
cloudpickle==2.0.0
cma==2.7.0
configparser==5.1.0
contextlib2==21.6.0
cycler==0.10.0
Cython==0.29.21
decorator==4.4.2
diGraph==0.1.4
docker-pycreds==0.4.0
docstring-parser==0.12
easydict==1.9
efficientnet-pytorch==0.6.3
einops==0.3.2
etelemetry==0.3.0
filelock==3.8.0
fire==0.4.0
flatbuffers==1.12
frozenlist==1.2.0
fsspec==2021.11.0
future==0.18.2
gast==0.2.2
gitdb==4.0.9
GitPython==3.1.24
google-auth==1.21.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
gpytorch==1.2.0
graphviz==0.14.2
grpcio==1.32.0
gym==0.21.0
h5py==2.10.0
histolab==0.4.0
huggingface-hub==0.10.0
hydra-core==1.1.1
idna==2.10
imageio==2.9.0
imgaug==0.4.0
importlib-metadata==4.9.0
importlib-resources==5.4.0
intel-openmp==2021.2.0
ipykernel==5.3.4
ipython==7.18.1
ipython-genutils==0.2.0
isodate==0.6.1
jedi==0.17.2
Jinja2==3.0.3
joblib==1.0.1
jsonargparse==4.0.0
jsonpatch==1.26
jsonpointer==2.0
jupyter-client==6.1.7
jupyter-core==4.6.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.1.0
kmeans1d==0.3.1
libtiff==0.4.2
llvmlite==0.38.0
lmdb==1.0.0
looseversion==1.0.3
lxml==4.9.2
Markdown==3.2.2
MarkupSafe==2.0.1
matplotlib==3.1.1
MedPy==0.4.0
mkl==2021.2.0
mkl-service==2.3.0
ml-collections==0.1.0
msgpack==1.0.1
msgpack-python==0.5.6
multidict==5.2.0
munch==2.5.0
networkx==2.5
nibabel==3.2.1
nipype==1.8.5
numba==0.55.0
numpy==1.18.5
nystrom-attention==0.0.11
oauthlib==3.1.0
olefile==0.46
omegaconf==2.1.1
onnx==1.8.0
onnx-tf==1.7.0
opencv-contrib-python==4.5.3.56
opencv-python==4.4.0.42
opencv-python-headless==4.5.3.56
openslide-python==1.1.2
opt-einsum==3.3.0
packaging==20.9
pandas==1.1.3
parso==0.7.1
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==6.2.1
plotly==5.4.0
pretrainedmodels==0.7.4
prettytable==1.0.0
progress==1.6
progressbar2==3.53.1
promise==2.3
prompt-toolkit==3.0.7
protobuf==3.13.0
prov==2.0.0
psutil==5.8.0
ptflops==0.6.3
ptyprocess==0.6.0
pyarrow==0.17.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.19
pydantic==1.8.2
pyDeprecate==0.3.1
pydicom==2.3.1
pydot==1.4.2
pyglet==1.5.21
Pygments==2.6.1
PYMIC==0.3.0
pymoo==0.4.2.2
pymop==0.2.4
pyparsing==2.4.5
PySocks==1.7.1
python-dateutil==2.8.1
python-utils==2.4.0
pytorch-ignite==0.4.1
pytorch-lightning==1.2.3
pytorch-toolbelt==0.4.0
pytz==2019.3
PyWavelets==1.1.1
pywt==1.0.6
PyYAML==5.3.1
pyzmq==19.0.2
rdflib==6.2.0
regex==2022.9.13
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
scikit-image==0.17.2
scikit-learn==1.0
scipy==1.7.2
seaborn==0.12.1
segmentation-models-pytorch==0.2.0
sentry-sdk==1.4.3
Shapely==1.7.1
shortuuid==1.0.8
SimpleITK==2.0.2
simplejson==3.18.3
six==1.15.0
sklearn==0.0
smmap==5.0.0
spams==2.6.5.4
stable-baselines==2.10.2
str2bool==1.1
subprocess32==3.5.4
tabulate==0.8.9
tbb==2021.2.0
tenacity==8.0.1
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.7.0
tensorboardX==2.1
tensorflow==1.15.0
tensorflow-addons==0.12.1
tensorflow-estimator==1.15.1
tensorflow-gpu==1.15.0
termcolor==1.1.0
thop==0.0.31.post2005241907
threadpoolctl==2.2.0
tifffile==2020.10.1
timm==0.4.12
tokenizers==0.12.1
topk==1.0
torch==1.9.0+cu111
torch-geometric==2.2.0
torchaudio==0.9.0
torchfile==0.1.0
torchmetrics==0.6.0
torchnet==0.0.4
torchstat==0.0.7
torchsummary==1.5.1
torchvision==0.10.0+cu111
torchviz==0.0.2
tornado==6.0.3
tqdm==4.48.2
traitlets==4.3.3
traits==6.3.2
transformers==4.20.1
ttach==0.0.3
typeguard==2.11.1
typing-extensions==3.7.4.3
umsgpack==0.1.0
urllib3==1.25.10
visdom==0.1.8.9
wandb==0.12.6
wcwidth==0.2.5
websocket-client==0.57.0
Werkzeug==1.0.1
wrapt==1.12.1
yacs==0.1.8
yarl==1.7.2
yaspin==2.1.0
zipp==3.1.0
